refactor shasum computation in an utility function
